flatpak: Explicitly specify the LibRaw build options
Currently the build detects the presence of the necessary dependencies
and automatically enables or disables the corresponding options. Unless
someone is keeping a close eye on the build logs, a change in the
manifest or the SDK might silently and unexpectedly change the build
options.

Ideally, the build would fail if a requested option can't be met, to
draw attention to this unforeseen change.

Unfortunately, the LibRaw build uses AC_MSG_WARN even for explicitly
enabled build options. This means that a missing dependency won't fail
the build even if the option was explicitly requested. While this is
not ideal, at least listing the options makes it clear what the
required outcome is. That way someone can manually check the logs from
time to time and see if it's living up to the requirements.

The set of build flags have been chosen to match the current reality of
automatically selected options. Hence, there shouldn't be any
user-visible change.

https://gitlab.gnome.org/GNOME/gnome-photos/merge_requests/91