gegl, test-gegl: Split out the code to convert a GeglBuffer to a format
This code is necessary whenever the validity of a floating point
GeglBuffer has to be asserted because floating point colour component
values are inherently not reproducible. Minor errors can creep in due
to differences in computer architecture, varying Babl conversions or
other changes in code. These errors don't indicate a bug and will
cause spurious test failures. Therefore, it's better to convert such
GeglBuffers into integer values before checksumming. eg., this is
relevant when a buffer is zoomed by a GeglSampler because the result
is floating point premultiplied alpha.

Subsequent commits will add a new codec API for GeglBuffer similar to
those for GdkPixbuf, which is going to offer on-the-fly zooming while
decoding a bitstream. Therefore, this will be useful for testing those
code paths.

https://gitlab.gnome.org/GNOME/gnome-photos/issues/63