src/photos-searchbar.c: Changed ctrl+a behaviour in search
Currently while searching photos by text, hitting the Ctrl+a key will
select all photos that matched the query.

This can lead to unwanted behaviour, like searching for a term then
hitting Ctrl+a Backspace in order to change the query term; this
delete all found images.

To fix this, changed the search bar code for key pressing event in
order to select the text instead of the photos.

https://gitlab.gnome.org/GNOME/gnome-photos/issues/124