build: Drop the Autotools build system
Meson has been the preferred build system for a while now - Build
Meta, Continuous, JHBuild, and the stable and nightly Flatpak
manifests have been building with Meson, and so has been Fedora.
So far, there are no known issues with it.

Having to maintain one less build system is a big win.

https://gitlab.gnome.org/GNOME/gnome-photos/merge_requests/113