tests: Isolate the test from any real user data
This uses the trackertestutils library to set up a temporary directory
for XDG_DATA_HOME and XDG_CACHE_HOME data, and spawn private Tracker
daemons which operate from this temporary directory. This means the
app under test will no longer change behaviour depending on data in
the HOME directory of the user running the test.