flatpak: Fix the grilo-plugins build
The grilo-plugins build has started to fallback to downloading grilo
as a subproject if certain dependencies aren't met [1]. In this case,
the grilo-pls-0.3 dependency is missing because it's disabled in the
grilo build via the -Denable-grl-pls=false option.

This is a problem because network access is not available in the build
sandbox.

[1] grilo-plugins commit 34d3c5844864f0d0
    https://gitlab.gnome.org/GNOME/grilo-plugins/-/merge_requests/68

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/124