flatpak: Fix the tracker and tracker-miners builds
Intltool was removed from version 3.34 of the GNOME SDK, but it's
still used by tracker-miners. Therefore it has to be bundled with the
build.

Tracker's master branches are now being used to develop version 3.0,
and it's not ready for use [1]. Hence the build has to be switched to
use the respective 2.3 branches.

Some changes by Debarshi Ray.

[1] https://discourse.gnome.org/t/tracker-has-branched-for-3-0-use-the-2-3-branch-if-you-want-stability/1889

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/118
https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/125