gegl: Add a version comparison macro for Babl
The SHA256 checksum of a GeglBuffer can be affected by changes in Babl
that alter the accuracy of the values generated by GEGL. These changes
are usually improvements to the existing fast paths in Babl, or the
addition or removal of fast paths; and the difference in the accuracy
of the values is too small to be perceived by the human eye, while the
performance benefits are significant.

Therefore, to avoid spurious failures, the test suite needs to support
multiple reference checksum strings - sometimes specific to a
particular Babl version. A subsequent commit will use this version
comparison macro to do so.