test-gegl: Support multiple reference checksums in /gegl/buffer/zoom/*
The SHA256 checksum of a GeglBuffer can be affected by changes in Babl
that alter the accuracy of the values generated by GEGL. Babl 0.1.67
added some AVX2 fast paths [1] that might occasionally affect the
checksum of GeglBuffers output by gegl:scale-ratio. The difference in
the accuracy of the values is too small to be perceived by the human
eye, while the performace benefits are significant.

Therefore, to avoid spurious failures, the test suite needs to support
multiple reference checksum strings.

[1] Babl commit 41a31deeb36a72bb

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/126