flatpak: Remove liboauth
liboauth is weakly maintained, and isn't actually used by GNOME Photos
at all. It was only included because libgdata used to have a hard
dependency on it. That's no longer the case.

Even though libgdata doesn't require liboauth by default, it's better
to explicitly specify it. It protects the build against any changes in
the SDK or the libgdata defaults.

https://gitlab.gnome.org/GNOME/gnome-photos/-/issues/116
https://gitlab.gnome.org/GNOME/libgdata/-/issues/1