test-pipeline: Test the deserialization of the Instagram filters
The Instagram filters make clever use of the "name" and "compat-name"
properties of GeglOperation and the PhotosOperationInstaPreset
enumerated type to prevent the Instagram filter names from appearing in
the XML, while still using them internally throughout the codebase.

Enumerated types can be vulnerable to backwards compatibility breakages
whenever a new name is added to the list, and it's necessary that older
XMLs continue to work with newer versions of the code.

Therefore, it's worth testing this.

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/134