Generate queries using SPARQL templates
The code to generate SPARQL queries was split across many different
source files, making it difficult to make big changes. Now the
queries are written out as templates and template substitution is used
to build the SPARQL that's sent to Tracker.

As a side effect, the SPARQL for the builtin "all" SearchType now uses
a UNION of the equivalent SPARQLs for collections and items, instead of
looking for all instances of rdfs:Resource and then using a FILTER to
weed out the irrelevant entries in the Tracker database.

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/135
https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/163