Switch to private instances of the Tracker 2.x daemons
GNOME 3.38 is switching to Tracker 3.x by default [1]. However, Photos
is not ready to switch this cycle.

It's desirable to avoid having two sets of Tracker daemons indexing
the same content as much as possible. One option is to start private
instances of the Tracker 2.x daemons tied to the lifetime of the GNOME
Photos application. Distributions can disable Tracker 2.x by default,
and it will only be activated when needed by Photos.

This also allows the Photos Flatpak to run Tracker 2.x daemons inside
the sandbox, ensuring it works on systems which don't have Tracker 2.x
available.

Based on code written by Yi-Soo An.

[1] https://gitlab.gnome.org/GNOME/Initiatives/-/issues/17

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/146