application: Rename a function for clarity
A subsequent commit will switch GNOME Photos to use private instances
of the Tracker 2.x daemons, where the daemons will be tied to the
lifetime of the application. This means that the application will be
responsible for also starting the miners for local content.

Therefore, it's necessary to clarify that this function only concerns
the online miners.

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/146