gegl: Don't exceed arbitrary maximum number of threads
GEGL has an arbitrary maximum value for GeglConfig:threads, currently
64, and exceeding it causes CRITCALs. This can make the build-time
tests fail on massively parallel hardware. For example, on a dual
POWER8 system used to build Debian ppc64el packages (2 CPUs, 10 cores
per CPU and 8 threads per core, for a total of 160 threads), we ask
GEGL to use 80 threads, which is more than 64.

The simple approach to this would be to clamp to the range
[1, GEGL_MAX_THREADS], but this would not be robust against possible
future changes to the value of GEGL_MAX_THREADS, so this commit
introspects the property to find out the true maximum at runtime.

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/154