Remove unnecessary {{item_where}} snippets
The {{item_where}} snippet is used to show the contents of a collection
in the COLLECTION_VIEW mode. This mode uses the "all" SPARQL template,
and therefore the {{item_where}} is unnecessary in all the other
templates.

A subsequent commit will port to Tracker 3, where the SPARQL templates
would have to aggregate metadata from different Tracker endpoints.
Since an item and its collection might belong to different endpoints,
not having the unused {{item_where}} snippets can greatly simplify the
SPARQL.

https://gitlab.gnome.org/GNOME/gnome-photos/-/issues/152