build, flatpak: Add an option for installing Flatpak-specific files
When running as a Flatpak, the application can use a private instance
of Tracker 3's filesystem miner on hosts without Tracker 3. This
requires a few extra files to be installed that are otherwise not
necessary.

This should make it easier for downstream distributors to ensure the
correctness of builds, without having to understand too many details.

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/174