base-item: Convert photos_base_item_trash to async API
This enables the caller to handle any error, and properly manages the
lifetime of the asynchronous operation by keeping the application and
the item alive while it's in flight.

A subsequent commit will use this to ensure that items marked for
deletion do get deleted when the main window is closed.

https://gitlab.gnome.org/GNOME/gnome-photos/-/issues/92