delete-notification: Ensure that items get deleted when app is quit
This ties the destruction of the DeleteNotification widget more tightly
to the deletion of the items represented by it. Instead of first
deleting the items separately and then destroying the widget, any
destruction, whether due to a timeout or the "close" button or the main
window being closed, will now delete the items.

The deletion of the items is guaranteed by keeping the application
alive while the operations are in flight.

https://gitlab.gnome.org/GNOME/gnome-photos/-/issues/92