thumbnail-factory: Simplify code
Tie the lifetime of the PhotosThumbnailerDBus proxy even more tightly
to that of the GDBusConnection, so that only one GMutex is enough.

As a nice side-effect, this addresses a race when handling errors from
GenerateThumbnail, if the GCancellable had been triggered in the
application and the thumbnailer process had crashed. Earlier, it was
possible for the GDBusConnection to have been destroyed and set to
NULL due to the crash before mutex_connection could be locked.

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/182