dlna-renderer: Don't fail photos_dlna_renderer_unshare without an error
The convention for GIO-style asynchronous APIs is to set an error when
returning a value that indicates a failure. In this case, returning
FALSE without setting an error would violate that.

There are some known exceptions to this convention, like
tracker_sparql_cursor_next, but those conditions don't apply here and
following the convention will avoid surprising the caller.

https://gitlab.gnome.org/GNOME/gnome-photos/-/merge_requests/196