tracker-queue: Add a second TrackerSparqlConnection for remote content
The user-generated metadata stored in ~/.local/share/gnome-photos needs
to be stored persistently because it's the canonical source for it.

However, the metadata for remote items merely needs to be cached, and
can be regenerated at ease from the corresponding online sources.
Therefore, it's worth using a separate database for it that will be
populated by the online miners. This will shield the user-generated
metadata from getting corrupted by a bug in the online miners.

https://gitlab.gnome.org/GNOME/gnome-photos/-/issues/83