source: Drop the "Downloads" folder from the list of local sources
The original intention was to look at the "Downloads" folder as a
temporary measure until we have a OS-wide solution for importing
specific files from the web that the user actually wants to keep for
later use, as opposed to files that the web browser randomly chose to
put in the "Downloads" folder. This can today be implemented through a
Flatpak portal.

However, this temporary solution is proving to be more of an annoyance
than anything. So it's better to take it out.

https://gitlab.gnome.org/GNOME/gnome-photos/-/issues/9